Create separate route folders
